+++ engine/src/main/java/dev/gamekit/ui/widgets/Column.java
@@ -1,32 +1,43 @@
 package dev.gamekit.ui.widgets;
 
 import dev.gamekit.ui.Container;
-import dev.gamekit.utils.Size;
-import dev.gamekit.utils.Reusable;
+import dev.gamekit.ui.Node;
+import dev.gamekit.utils.Constraints;
+
+import java.util.List;
 
 /** Column container arranges its children horizontally */
 public class Column extends Container {
   @Override
-  protected Size getContentSize() {
-    int widestWidth = 0;
-    int combinedHeight = 0;
-
-    for (var child : children) {
-      widestWidth = Math.max(widestWidth, child.getComputedSize().width);
-      combinedHeight += child.getComputedSize().height;
-    }
-
-    Reusable.SIZE.set(widestWidth, combinedHeight);
-    return Reusable.SIZE;
+  protected List<Node> getChildren() {
+    return List.of();
   }
 
   @Override
-  protected void updateContentPositions() {
-    int ypos = 0;
+  public void onLayout(Constraints constraints) {
 
-    for (var child : children) {
-      child.getComputedPosition().set(0, ypos);
-      ypos += child.getComputedSize().height;
-    }
   }
+  //  @Override
+  //  protected Size getContentSize() {
+  //    int widestWidth = 0;
+  //    int combinedHeight = 0;
+  //
+  //    for (var child : children) {
+  //      widestWidth = Math.max(widestWidth, child.getComputedSize().width);
+  //      combinedHeight += child.getComputedSize().height;
+  //    }
+  //
+  //    Reusable.SIZE.set(widestWidth, combinedHeight);
+  //    return Reusable.SIZE;
+  //  }
+  //
+  //  @Override
+  //  protected void updateContentPositions() {
+  //    int ypos = 0;
+  //
+  //    for (var child : children) {
+  //      child.getComputedPosition().set(0, ypos);
+  //      ypos += child.getComputedSize().height;
+  //    }
+  //  }
 }