+++ engine/src/main/java/dev/gamekit/ui/widgets/Row.java
@@ -1,32 +1,44 @@
 package dev.gamekit.ui.widgets;
 
 import dev.gamekit.ui.Container;
-import dev.gamekit.utils.Size;
-import dev.gamekit.utils.Reusable;
+import dev.gamekit.ui.Node;
+import dev.gamekit.utils.Constraints;
+
+import java.util.List;
 
 /** Column container arranges its children vertically */
 public class Row extends Container {
   @Override
-  protected Size getContentSize() {
-    int combinedWidth = 0;
-    int tallestHeight = 0;
-
-    for (var child : children) {
-      combinedWidth += child.getComputedSize().width;
-      tallestHeight = Math.max(tallestHeight, child.getComputedSize().height);
-    }
-
-    Reusable.SIZE.set(combinedWidth, tallestHeight);
-    return Reusable.SIZE;
+  protected List<Node> getChildren() {
+    return List.of();
   }
 
   @Override
-  protected void updateContentPositions() {
-    int xpos = 0;
+  public void onLayout(Constraints constraints) {
 
-    for (var child : children) {
-      child.getComputedPosition().set(xpos, 0);
-      xpos += child.getComputedSize().width;
-    }
   }
+
+  //  @Override
+  //  protected Size getContentSize() {
+  //    int combinedWidth = 0;
+  //    int tallestHeight = 0;
+  //
+  //    for (var child : children) {
+  //      combinedWidth += child.getComputedSize().width;
+  //      tallestHeight = Math.max(tallestHeight, child.getComputedSize().height);
+  //    }
+  //
+  //    Reusable.SIZE.set(combinedWidth, tallestHeight);
+  //    return Reusable.SIZE;
+  //  }
+  //
+  //  @Override
+  //  protected void updateContentPositions() {
+  //    int xpos = 0;
+  //
+  //    for (var child : children) {
+  //      child.getComputedPosition().set(xpos, 0);
+  //      xpos += child.getComputedSize().width;
+  //    }
+  //  }
 }